fix: correct 'this' binding issue in setInterval function call
This commit fixes a critical issue where the server was breaking due to improper 'this' binding within a setInterval function call. Previously, the .run() method was being invoked without the correct context, causing it to lose reference to the class instance and failing to recognize the this.getStatus() method within its context.

The solution involves adding a .bind() to ensure that the .run() method retains the proper 'this' reference, as shown below:

```javascript
setInterval(
  () => {
    this._functions.timed[functionName].run.bind(
      this._functions.timed[functionName],
    )();
  },
  this._moment
    .duration(
      +`${functionDefinition.intervalTime}`,
      `${functionDefinition.intervalMeasure}`,
    )
    .as('milliseconds'),
);